+++ back/server/src/main/kotlin/com/metamong/infra/persistence/apartment/projection/ApartmentComplexListProjection.kt
@@ -3,7 +3,7 @@ package com.metamong.infra.persistence.apartment.projection
 data class ApartmentComplexListProjection(
     val id: Long,
     val name: String,
-    val builtYear: Short?,
+    val builtYear: Int?,
     val totalHousehold: Int?,
     val eupmyeondongRiCode: Int?,
     val addressJibun: String?,
